app/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.5
psycopg2-binary>=2.9
scikit-learn>=1.3
joblib>=1.3
python-dotenv>=1.0
pytest>=7.4
pytest-cov>=4.1
httpx>=0.25
pandas>=2.0
numpy>=1.24
sqlalchemy>=2.0
xgboost>=3.1
optuna>=3.0
optuna-integration